OESjava-Core0: refactoring, bug fixed
* Decoupling of communication and execution components
*  Double values in statistics
* Possibility to log all active objects

+++ Java/Core0/src/main/java/de/oes/core0/inventorymanagment/DailyDemand.java
@@ -52,7 +52,7 @@ public class DailyDemand extends ExogenousEvent {
 	public DailyDemand createNextEvent() {
 		return new DailyDemand(
 				this.getSim(), 
-				this.getOccTime() +  this.reccurence(), 
+				this.getOccTime().longValue() +  this.reccurence(), 
 				DailyDemand.quantity(), 
 				this.shop);
 	}
